+++ docs/Whats-New.md
@@ -4,6 +4,10 @@ This page lists the history of changes across stable Phobos releases and also al
 
 ## Migrating
 
+```{hint}
+You can use the migration utility to apply most of the changes automatically via a corresponding sed script file. To use the utility, right-click on `utilities/MigrationUtility.ps1`, select `Run with PowerShell` and select the needed files.
+```
+
 ### From vanilla
 
 - SHP debris hardcoded shadows now respect `Shadow=no` tag value, and due to it being the default value they wouldn't have hardcoded shadows anymore by default. Override this by specifying `Shadow=yes` for SHP debris.
@@ -360,6 +364,9 @@ Phobos fixes:
 - Fixed interceptors intercepting projectiles fired by friendly objects if the said object died after firing the projectile (by Starkku)
 - Fixed interceptor weapons with `Inviso=true` projectiles detonating the projectile at wrong coordinates (by Starkku)
 - Fixed some possible configuration reading issues when using Phobos with patches that rename `uimd.ini` (by Belonit)
+
+Non-DLL:
+- Implemented a tool (sed wrapper) to semi-automatically upgrade INIs to use latest Phobos tags (by Kerbiter)
 </details>